remove League\Uri\Factory class
