--- .github/ISSUE_TEMPLATE.md
@@ -1,22 +1,0 @@
-## Issue summary
-
-_(Please explain in plain english your issue)_
-
-### System informations
-
-_(In case of a bug report Please complete the table below)_
-
-| Information | Description |
-|--------------|---------|
-| League\Uri\Schemes version |  |
-| PHP/HHVM version |  |
-| OS Platform |  |
-
-
-## Standalone code, or other way to reproduce the problem
-
-_(Please complete the text below to help us fix the issue)_
-
-### Expected result
-
-### Actual result
