+++ .github/ISSUE_TEMPLATE/Bug_Report.md
@@ -1,0 +1,30 @@
+---
+name: 🐛 Bug Report
+about: Something is broken? 🔨
+---
+
+### Bug Report
+
+_(Fill in the relevant information below to help triage your issue.)_
+
+| Information | Description |
+|--------------|---------|
+| Version |  |
+| PHP version |  |
+| OS Platform |  |
+
+#### Summary
+
+_(Please explain in plain english your bug)_
+
+#### Standalone code, or other way to reproduce the problem
+
+_(Please complete the text below to help us fix the issue)_
+
+#### Expected result
+
+_(What was the expected (correct) behavior?)_
+
+#### Actual result
+
+_(What is the current (buggy) behavior?)_
