+++ .github/ISSUE_TEMPLATE/Feature_Request.md
@@ -1,0 +1,16 @@
+---
+name: 🎉 Feature Request
+about: Do you have a new feature in mind?
+---
+
+### Feature Request
+
+|    Q        |   A
+|------------ | ------
+| New Feature | yes
+| BC Break    | yes/no
+
+
+#### Proposal
+
+_(Please explain in plain english the feature you would like to see implemented.)_
