+++ .github/ISSUE_TEMPLATE/Question.md
@@ -1,0 +1,14 @@
+---
+name: ❓ Question
+about: Are you unsure about something?
+---
+
+_(Fill in the relevant information below to help triage your issue.)_
+
+|    Q        |   A
+|------------ | ------
+| Version     | x.y.z
+
+### Question
+
+_(Please explain in plain english your question)_
