--- .github/PULL_REQUEST_TEMPLATE.md
@@ -1,25 +1,0 @@
-## Introduction
-
-_(Please explain the library current status regarding your PR.)_
-
-## Proposal
-
-### Describe the new/upated/fixed feature
-
-_(Please explain your proposal. If this is a issue fix, please provide the issue number. If you have not yet created an issue, please do so now)_
-
-### Backward Incompatible Changes
-
-_(Describe if they exists BC break)_
-
-### Targeted release version
-
-_(Indicate the release version targeted for your PR)_
-
-### PR Impact
-
-_(Describe the PR impact on the current public API)_
-
-## Open issues
-
-_(Describe possible open issues and/or future scope if any depending on this RFC acceptance)_
