+++ .github/PULL_REQUEST_TEMPLATE/Bug_Fix.md
@@ -1,0 +1,18 @@
+---
+name: 🐛 Bug Report
+about: Something is broken and you have a failing Unit Test? 🔨
+---
+
+### Bug Report
+
+_(Fill in the relevant information below to help triage your issue.)_
+
+| Information | Description |
+|--------------|---------|
+| Version |  |
+| PHP version |  |
+| OS Platform |  |
+
+#### Summary
+
+_(Please explain in plain english the failing test)_
