+++ .github/PULL_REQUEST_TEMPLATE/New_Feature.md
@@ -1,0 +1,26 @@
+---
+name: 🎉 New Feature
+about: You have implemented some neat idea that you want to make part of **League\Uri\Components** ? 🎩
+---
+
+<!--
+Thank you for submitting new feature!
+Pick the target branch based according to these criteria:
+  * submitting a bugfix: target the lowest active stable branch: 2.6
+  * submitting a new feature: target the master branch.
+  * submitting a BC-breaking change: target the master branch
+-->
+
+### New Feature
+
+_(Fill in the relevant information below to help triage your issue.)_
+
+|    Q        |   A
+|------------ | ------
+| New Feature | yes/no
+| BC Break    | yes/no
+
+
+#### Proposal
+
+_(Please explain in plain english the feature you would like to see implemented.)_
