+++ .php_cs
@@ -6,7 +6,7 @@ League.Uri (http://uri.thephpleague.com)
 @package    League\Uri
 @subpackage League\Uri\Schemes
 @author     Ignace Nyamagana Butera <nyamsprod@gmail.com>
-@license    https://github.com/thephpleague/uri-components/blob/master/LICENSE (MIT License)
+@license    https://github.com/thephpleague/uri-schemes/blob/master/LICENSE (MIT License)
 @version    2.0.0
 @link       https://github.com/thephpleague/uri-schemes
